+++ internal/adapter/handlebars/handlebars_test.go
@@ -137,6 +137,14 @@ func TestConcatHelper(t *testing.T) {
 	testString(t, "{{concat '> ' 'A quote'}}", nil, "> A quote")
 }
 
+func TestSubstringHelper(t *testing.T) {
+	testString(t, "{{substring '' 2 4}}", nil, "")
+	testString(t, "{{substring 'A full quote' 2 4}}", nil, "full")
+	testString(t, "{{substring 'A full quote' 40 4}}", nil, "")
+	testString(t, "{{substring 'A full quote' -5 5}}", nil, "quote")
+	testString(t, "{{substring 'A full quote' -5 6}}", nil, "quote")
+}
+
 func TestJoinHelper(t *testing.T) {
 	test := func(items []string, expected string) {
 		context := map[string]interface{}{"items": items}
@@ -273,8 +281,8 @@ func testLoader(opts LoaderOpts) *Loader {
 	loader.RegisterHelper("style", helpers.NewStyleHelper(opts.Styler, &util.NullLogger))
 	loader.RegisterHelper("slug", helpers.NewSlugHelper("en", &util.NullLogger))
 
-	formatter := func(path, title string) (string, error) {
-		return path + " - " + title, nil
+	formatter := func(context core.LinkFormatterContext) (string, error) {
+		return context.Path + " - " + context.Title, nil
 	}
 	loader.RegisterHelper("format-link", helpers.NewLinkHelper(formatter, &util.NullLogger))